api/requirements.txt
APScheduler==3.10.4
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
playwright==1.43.0
psycopg2-binary==2.9.9
pyee==11.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.2
wsproto==1.2.0
